Update wp-config.php to support Docker environment
This commit introduces modifications to the wp-config.php to allow WordPress to read environment variables when running in a Docker container. Database settings, unique keys, and salts are now defined by using environment variables, improving the scalability and portability of the WordPress installation. Additionally, checks for the existence of a proxy server and whether HTTPS is being used have also been added.